repo: drop classmethod and superfluous parameter
AptKeyManager.is_key_installed() was a classmethod and had a parameter
with the path to search. However, the function is only ever called with
an instance of the class, and then the parameter is misleading because
the instance already has its own keyring path. Now AptKeyManager
instances will look for keys inside their own _keyrings_path.