repo: install package repositories in "arbitrary" paths (#80)
This commit adds a new public function, "install_in_root()", that
installs package repositories in a location that contains an Apt-based
system, and _not_ the host system. The intention is that eventually
said system will be chroot'ed into, at which point an "apt update" will
pick up the new repositories.

The motivation is to support package repositories in overlays in
Rockcraft.