keys: make AptKeyManager.install_key() more robust
This commit relaxes the restriction of the provided key contents having
only a single (primary) key, and makes AptKeyManager better handle the
fact that "gpg import" might print out useful diagnostics messages even
if the key contents are imported "successfully".

To do this, we capture gpg's output and add a new optional parameter
for the desired "target" key fingerprint (the key that we desire to
import even if other keys exist in the contents). This fingerprint is
used when naming the imported file, and then:

1) If we don't have a desired fingerprint, proceed as before.
2) If we have a desired fingerprint, check for it in the imported file,
  failing if it's not there.

The motivation for this change is that there are keys "in the wild" that
generate gpg error messages but still contain a valid, non-expired
desired key. This was supported before but regressed when we moved away
from apt-key.

This commit is the second one on the road to fixing the "multiple keys
in the same file" regression.