keys: make AptKeyManager.install_key() more robust
This commit relaxes the restriction of the provided key contents having
only a single (primary) key, and makes AptKeyManager better handle the
fact that "gpg import" might fail (a non-zero return code) for many
reasons while still importing the desired key successfully.

To do this, we capture gpg's output and add a new optional parameter
for the desired "target" key fingerprint (the key that we desire to
import even if other keys exist in the contents). Then:

If gpg finishes successfully:

1) If we don't have a desired fingerprint, proceed as before.
2) If we have a desired fingerprint, check for it in the imported file.

If gpg returns a non-zero code:

3) If we don't have a desired fingerprint, raise an error with the
  captured output.
4) If we have a desired fingerprint, check for it in the imported file.
  If found, log the captured output but don't raise errors.

The motivation for this change is that there are keys "in the wild" that
generate gpg errors but still contain a valid, non-expired desired key.
This was supported before but regressed when we moved away from apt-key.