keys: forward the key-id to install_key()
The final piece to support multiple keys is to forward the key-id from
the PackageRepositoryApt model to the AptKeyManager.install_key() call.

This is necessary because now that we support key contents containing
multiple keys we must name the imported key itself in a deterministic
useful way - the previous behavior of using the "first fingerprint" is
not appropriate because the fingerprint might refer to an expired or
otherwise undesirable key. We can use the model's key-id because that
is provided by the user itself, and is also used in naming the file
in the assets dir.