sources: only add "compatible" architectures
This commit both fixes the order of the parameters in the call to
"dpkg --add-architecture" (the parser is picky) and updates the logic
to only make said call if the target architecture (that is, the one
coming from the package-repository definition) is "compatible" with the
host architecture (that is, the architecture of the system running the
code).

This is necessary because incompatible mixes typically break the system.
For example, adding "armhf" on an "amd64" system breaks subsequent calls
to "apt update" because apt will try to fetch package listings for the
"armhf" architecture on the default Ubuntu archives, which don't have
them.

Therefore, only the following pairs result in adding the target arch:

- host amd64, target i386;
- host arm64, target armhf.