feat: add needed fields for imagecraft (#120)
* feature: Add a key-id field to PackageRepositoryAptPPA

Adds a key-id field to PackageRepositoryAptPPA to let users define
the key they want and not have to get it from the server.

Adding this field also helps being reducing additions to this class
in Imagecraft.

Signed-off-by: Paul Mars <paul.mars@canonical.com>

* docs: Document the priority field

Adds missing documentation of the priority field. Helps reducing
the delta with the documentation visible on snapcraft.io.

Signed-off-by: Paul Mars <paul.mars@canonical.com>

* feat: Make sure components are unique

Makes sure the provided components lists contains only unique elements.

Signed-off-by: Paul Mars <paul.mars@canonical.com>

* feat: Add series and pocket fields to PackageRepositoryApt

This is an alternative to the suites field and enable properly
splitting this value (and to avoid parsing). Series/pocket cannot
be used to at the same time as suites.

* fix: Use a enum for UCA pockets

Signed-off-by: Paul Mars <paul.mars@canonical.com>

* feat: use series/pocket to generate deb822 sources list

When the pocket if defined, generate a list of suites from the series and the pocket.
_get_suites is tailored to what is expected in an official ubuntu system.

Signed-off-by: Paul Mars <paul.mars@canonical.com>

---------

Signed-off-by: Paul Mars <paul.mars@canonical.com>