fix: handle existing ubuntu deb822 sources (#136)
The problem is this: since Noble, the default Ubuntu archives are listed as
deb822 sources in /etc/apt/sources.lists.d/ubuntu.sources as signed by a key in
/usr/share/keyrings. Trying to declare a package-repository with the url
of an official archive (e.g. to add an architecture) would then fail because
apt complained that a given suite under archive.ubuntu.com/ubuntu was being
declared as Signed-By two keyrings (the official one and the craft-added one).

This commit updates the handling of sources for the specific case of those
sources that are also declared in ubuntu.sources:
- IF a package-repository has an url that matches an official repo,
- AND IF this repo is already declared in a deb822-format sources file,
- AND IF the requested key-id is present in the existing sources' declared
  keyring,

... then the new sources file will point to the existing keyring, instead
of the usual CRAFT-xxx keyring. This commit updates a lot of code, but most of
of it is moving existing code around, and the new code specifically only
applies to the 'ubuntu.sources' file. This is to minimize the blast-radius of
this change and allow this to be a 'fix' rather than a 'feat'.

Fixes #129

---------

Co-authored-by: Callahan <callahan.kovacs@canonical.com>