fix: retry key receiving on port TCP/80 if using default keyserver (#189)
By default gpg uses the hkp protocol (over HTTP) on port TCP/11371. This port may be filtered in some environments and may not be accepted by proxies. When gpg timeout trying to connect to the default keyserver, retry with on port TCP/80. The default protocol is unchanged but must be explicitly set when setting a port.

Also fix AptGPGKeyInstallError doing unnecessary work since moving from apt-key to gpg.

---------

Signed-off-by: Paul Mars <paul.mars@canonical.com>