feat: add ability to set default releases to old-releases (#192)
This adds a function at `craft_archives.defaults.use_old_releases`.
When called, it checks the given root's `etc/apt/sources.list` file or
`/etc/apt/sources.list.d/ubuntu.sources` for sources that are available
on `old-releases.ubuntu.com`. For each source it finds there, it swaps
the source to use `old-releases.ubuntu.com`.

While this is safe to run on any release, it is intended only to run on
EOL Ubuntu releases.

CRAFT-4634

Signed-off-by: Alex Lowe <alex.lowe@canonical.com>
Co-authored-by: JJ Coldiron <jj.coldiron@canonical.com>